requirements.txt
streamlit
openai
langchain
faiss-cpu
python-dotenv
semantic-text-splitter
sentence-transformers
rank_bm25
unstructured[pdf]
langchain
faiss-cpu
huggingface-hub
docx2txt







